Support personal clones of subprojects
When working on AsteroidOS recipes, it's convenient to have both a fork
of the subproject and also to set the upstream url.  However, the
prepare-build.sh file would then only retrieve updates from the fork
rather than upstream.  This addresses that issue by checking whether
there is an upstream or not; if not, it does the fork as it always had.
If there is an upstream, it pulls the named head from that fork instead,
making it easy and reliable to maintain one's own fork for contributions
back to the main project.

Signed-off-by: Ed Beroset <beroset@ieee.org>