Update layer configuration whenever a new layer is added.
Previously one could provide a new platform to the `devices` variable. This implies a new layer under `src/meta-smartwatch/meta-*`.
But the build configuration (bblayers.conf) is never updated with this new information as that file may already exists.

This adds support for updating the layer configuration whenever a new layer is added to the `layer_conf` variable and or a new layer becomes available in the `meta-smartwatch` repository.

Signed-off-by: Darrel Griët <dgriet@gmail.com>