Do not generate an empty bblayers.conf file upon update
https://github.com/AsteroidOS/asteroid/commit/46339cad318aeb7e364377e7c55f16d5f20943e2
introduced behaviour such that the 'current' state of `bblayers.conf` is copied to
a temporary file layers are updated and this bblayers.conf is overwritten with
the temporary file generated by `mktemp`.

However, this introduces an issue where an empty `bblayers.conf` is created if
the user decides to update layers perior to building (i.e. do
`./prepare-build.sh update` first) as the temporary file (created with `mktemp`)
is copied without any contents because `bblayers.conf` doesn't exist yet.

This empty file doesn't contain the `layer_line` pattern, so nothing will be
appended.

To fix this issue, do not generate an empty `bblayers.conf` file upon update if
it doesn't exist.

Signed-off-by: Darrel Griët <dgriet@gmail.com>